Change to non-insane keycodes for plover-hid
Previously we had an idea to stick to the geminipr keycodes as far as
possible, but that only made implementing this protocol much harder than
it needed to be, so switch to a sequential numbering scheme, and add all
keys outside of the standard steno keys as `X`-keys instead.

The ordering of the standard steno keys now corresponds to the ordering
of the TXBolt protocol.